+++ changelog.md
@@ -1,3 +1,6 @@
+### 9/13/2023 0.4.0-alpha.2
+- fix issue of unicam not moving when Cascade shadow map used
+- unicam speed increases up when shft key is pressed
 ### 9/13/2023 0.4.0-alpha.1
 - azimuth and elevation algorithm changed 
 ### 9/12/2023 0.4.0-alpha.0